int -> size_t

+++ src/data_structure/segment_tree.cc
@@ -16,7 +16,7 @@ class SegmentTree {
       : identity_element_(identity_element), data_(data), function_(function) {
     Build();
   }
-  void Update(int position, Monoid new_value) {
+  void Update(size_t position, Monoid new_value) {
     position += n_;
     node_[position] = new_value;
     while (position > 0) {
@@ -25,7 +25,7 @@ class SegmentTree {
     }
   }
 
-  Monoid Query(int left, int right) {
+  Monoid Query(size_t left,size_t right) {
     Monoid vl = identity_element_, vr = identity_element_;
     for (left += n_, right += n_; left < right; left >>= 1, right >>= 1) {
       if (left & 1) vl = function_(vl, node_[left++]);
@@ -34,25 +34,25 @@ class SegmentTree {
     return function_(vl, vr);
   }
 
-  Monoid operator[](int position) const { return node_[n_ + position]; }
+  Monoid operator[](size_t position) const { return node_[n_ + position]; }
 
  private:
-  int n_;
+  size_t n_;
   Monoid identity_element_;
   std::vector<Monoid> data_;
   std::vector<Monoid> node_;
   Function function_;
   void Build() {
-    int SIZE = data_.size();
+    size_t SIZE = data_.size();
     n_ = 1;
     while (n_ < SIZE) {
       n_ <<= 1;
     }
     node_.assign(2 * n_, identity_element_);
-    for (int i = 0; i < SIZE; ++i) {
+    for (size_t i = 0; i < SIZE; ++i) {
       node_[i + n_] = data_[i];
     }
-    for (int i = n_ - 1; i > 0; --i) {
+    for (size_t i = n_ - 1; i > 0; --i) {
       node_[i] = function_(node_[2 * i], node_[2 * i + 1]);
     }
   }
